Disassemble createApp
Place factory-factories in its own module, as well as the module ID resolution
helper.

Reorder various declarations to hopefully be slightly more sensible.
